VF-9 - he's walking on a sphere - however...
Much of this is still fucked. In particular, he seems to glitch out
if he travels too far. It is not clear to me why this might be.

It's possible this is physics acting - a rectangle does not have a
stable equilibrium standing up. Better might well be something like
an egg shape.